ci: enable 3 more clippy lints
Enable these lints in the xclippy/xclippy-fix aliases:
- clippy::needless_continue
- trivial_numeric_casts
- unused_lifetimes

Fix the warnings these lints caught in the codebase:
- Remove redundant continue expressions at end of match arms and loops
- Remove unnecessary numeric casts (e.g., u64 as u64)
- Allow unused_lifetimes in generated lalrpop parser code

+++ miden-vm/tests/integration/operations/u32_ops/arithmetic_ops.rs
@@ -468,7 +468,7 @@ fn u32widening_mul() {
     let a = rand_value::<u32>();
     let b = rand_value::<u32>();
     let result = a as u64 * b as u64;
-    let lo = (result % U32_BOUND) as u64;
+    let lo = result % U32_BOUND;
     let hi = result / U32_BOUND;
     let test = build_op_test!(asm_op, &[b as u64, a as u64]);
     test.expect_stack(&[lo, hi]);
